version3.5.71
修改部分bug

+++ userCode/tasks/Inc/AutoTask.h
@@ -19,7 +19,7 @@ void ClawTask();
 
 /*外部函数声明-------------------------------------------------------------*/
 extern void AutoChassisStop();//Realized in ChassisTask
-extern void AutoChassisSet(uint16_t x,uint16_t y,uint16_t o);//Realized in ChassisTask
+extern void AutoChassisSet(float x, float y, float o);//Realized in ChassisTask
 extern void AutoArmSet(float armzPos,float arm1Pos,float arm2Pos);//Realized in ArmTask
 extern void AutoTraySet(uint8_t trayflag);//Realized in ArmTask
 extern void AutoClawSet(uint8_t clawflag);//Realized in ServoTask
